% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_session_info.R
\name{child_session_info}
\alias{child_session_info}
\title{An Rmd child template that produces a collapsed session info box}
\usage{
child_session_info(pkgs = c("loaded", "attached", "installed")[1])
}
\arguments{
\item{pkgs}{Which packages to show. It may be:
\itemize{
\item \code{NULL} or \code{"loaded"}: show all loaded packages,
\item \code{"attached"}: show all attached packages,
\item \code{"installed"}: show all installed packages,
\item a character vector of package names. Their (hard) dependencies are
also shown by default, see the \code{dependencies} argument.
}}
}
\description{
When called in an Rmd document this function generates child document containing an expandable
session info box.
}
\section{Examples}{


Call this function on a separate line using inline code \verb{\{r, eval = FALSE, echo = TRUE\}r child_session_info()}.
}

\section{css}{

In CSS Use the \code{.session} class to target the whole expandable box.
Use \code{summary.session-header} to target the box header.
Use \code{.session-details} to target the content of the box.
}

