% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugodownplus-package.R
\docType{package}
\name{hugodownplus-package}
\alias{hugodownplus}
\alias{hugodownplus-package}
\title{hugodownplus: Extend 'hugodown's markdown document with toc and html boxes.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

hugodownplus extends 'hugodown's md_document with toc and HTML boxes. It allows the use of the arguments 'includes' and 'pandoc_args' which are available in rmarkdown::md_document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimTeaFan/hugodownplus}
  \item Report bugs at \url{https://github.com/TimTeaFan/hugodownplus/issues}
}

}
\author{
\strong{Maintainer}: Tim Tiefenbach \email{mailme@tim-tiefenbach.de} (\href{https://orcid.org/0000-0001-9443-2434}{ORCID})

Authors:
\itemize{
  \item Shafayet Khan Shafee \email{sshafee@isrt.ac.bd}
}

}
\keyword{internal}
