% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_session_info.R
\name{child_session_info}
\alias{child_session_info}
\title{An Rmd child template that produces a collapsed session info box}
\usage{
child_session_info()
}
\description{
An Rmd child template that produces a collapsed session info box
}
\section{Examples}{

}

