% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugodownplus-package.R
\docType{package}
\name{hugodownplus-package}
\alias{hugodownplus}
\alias{hugodownplus-package}
\title{hugodownplus: Extend 'hugodown's markdown document with toc and pandoc arguments.}
\description{
hugodownplus extends 'hugodown' md_document with toc and pandoc arguments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimTeaFan/hugodownplus}
  \item Report bugs at \url{https://github.com/TimTeaFan/hugodownplus/issues}
}

}
\author{
\strong{Maintainer}: Tim Tiefenbach \email{mailme@tim-tiefenbach.de} (\href{https://orcid.org/0000-0001-9443-2434}{ORCID})

}
\keyword{internal}
